+++ tests/Stubs/StubAppend.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * Author       : Samuel Dervis
+ * Email        : sam@samerior.com
+ * Company      : Samerior Group
+ */
+
+namespace Samerior\LaravelSidebar\Tests\Stubs;
+
+use Samerior\LaravelSidebar\Library\Core\DefaultAppend;
+
+/**
+ * Class StubAppend
+ * @package Samerior\LaravelSidebar\Tests\Stubs
+ */
+class StubAppend extends DefaultAppend
+{
+
+}
