+++ tests/Stubs/StubBadge.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * Author       : Samuel Dervis
+ * Email        : sam@samerior.com
+ * Company      : Samerior Group
+ */
+
+namespace Samerior\LaravelSidebar\Tests\Stubs;
+
+use Samerior\LaravelSidebar\Library\Core\DefaultBadge;
+
+/**
+ * Class StubBadge
+ * @package Samerior\LaravelSidebar\Tests\Stubs
+ */
+class StubBadge extends DefaultBadge
+{
+
+}
