+++ tests/Stubs/StubGroup.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * Author       : Samuel Dervis
+ * Email        : sam@samerior.com
+ * Company      : Samerior Group
+ */
+
+namespace Samerior\LaravelSidebar\Tests\Stubs;
+
+use Samerior\LaravelSidebar\Library\Core\DefaultGroup;
+
+/**
+ * Class StubGroup
+ * @package Samerior\LaravelSidebar\Tests\Stubs
+ */
+class StubGroup extends DefaultGroup
+{
+
+}
