+++ tests/Stubs/StubMenu.php
@@ -1,0 +1,20 @@
+<?php
+/**
+ * Author       : Samuel Dervis
+ * Email        : sam@samerior.com
+ * Company      : Samerior Group
+ */
+
+namespace Samerior\LaravelSidebar\Tests\Stubs;
+
+
+use Samerior\LaravelSidebar\Library\Core\DefaultMenu;
+
+/**
+ * Class StubMenu
+ * @package Samerior\LaravelSidebar\Tests\Stubs
+ */
+class StubMenu extends DefaultMenu
+{
+
+}
